+++ crates/utils/database/src/lib.rs
@@ -37,6 +37,7 @@ use media_models::{
     ShowSpecifics, VideoGameSpecifics, VisualNovelSpecifics,
 };
 use migrations::AliasedCollectionToEntity;
+use rust_decimal::Decimal;
 use rust_decimal::prelude::ToPrimitive;
 use rust_decimal_macros::dec;
 use sea_orm::{
@@ -488,7 +489,7 @@ pub async fn calculate_user_activities_and_summary(
         shows: HashSet<String>,
         show_seasons: HashSet<i32>,
         anime_episodes: HashSet<i32>,
-        manga_chapters: HashSet<i32>,
+        manga_chapters: HashSet<Decimal>,
         manga_volumes: HashSet<i32>,
     }
     type Tracker = HashMap<Date, TrackerItem>;
